+++ kroxylicious/src/main/java/io/kroxylicious/proxy/micrometer/MicrometerConfigurationHook.java
@@ -7,7 +7,19 @@ package io.kroxylicious.proxy.micrometer;
 
 import io.micrometer.core.instrument.MeterRegistry;
 
-public interface MicrometerConfigurationHook {
+public interface MicrometerConfigurationHook extends AutoCloseable {
+    /**
+     * Configures this hook into the given registry.
+     *
+     * @param targetRegistry meter registry.
+     */
     void configure(MeterRegistry targetRegistry);
+
+    /**
+     * Releases resources that may have been registered by this hook.  Implementations
+     * of this method must not throw unchecked exceptions.
+     */
+    default void close() {
+    }
 
 }